refactor: moved page logic to /page package
This allows for the page rendering and template logic to be better
abstracted from the main application code. Additionally, `page.go`
contains this line of code that is problematic:

    //go:embed template.html
    var tmpl string

This code is problematic because it creates a global variable where
there should be a compile-time constant. Thus, creating a new package
limits the scope of this variable somewhat.

Signed-off-by: Ben Raz <ben.raz2008@gmail.com>