feat: comprehensive rewrite of server
The server logic was first moved to it's own package. The reason for
this is that the code for managing the union file system gets pretty
large and it makes more sense to have it confined to it's own package.

The union file system was implemented according to these rules:

1. Any number of directories (called roots) can be broadcast from.
2. If a request is made to a directory all files will be returned from
   all roots that have that directory.
3. If a request is made to a file, the first root that has that
   file will take precedence over all roots, whether they have the file
   or not.
4. If a request is made whose path could be either a file or a directory
   depending on which root it is rooted under, the file will take
   precedence (example: the request is `/hello` and root `/a` has
   `/a/hello/` and root `/b` has `/b/hello`. `/b/hello` will 'win' in
   this circumstance). If the request ends with a `/`, the directory
   will take precedence.

For more, see function `server.resolve()`

Additionally, because of the latter part of rule 4, page/template.html
was modified to end all links to directories with a `/` to avoid
unexpected behavior.

Signed-off-by: Ben Raz <ben.raz2008@gmail.com>